third_party/blink/web_tests/platform/ios/fullscreen/full-screen-with-flex-item-expected.txt

Enter Fullscreen
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenchange)
EXPECTED (735 == '735') OK
END OF TEST
